pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "youtube_music_blend"
version = "0.0.0"
requires-python = ">=3.10"
dependencies = [
    "ytmusicapi",
    "pyyaml",
]

[project.scripts]
ytmb = "ytmb.__main__:main"

[tool.hatch.build.targets.wheel]
packages = ["ytmb"]
